Better errors for large record types
Before, we just went until any field caused a problem. If it was an
early field, there may be 10 more that all match, but have not been
unified. The result was that the *looked* different in the type diff,
but it was all visual noise.

So this version unifies all fields no matter what. In simple cases,
this makes the type errors much more specific. I can imagine it getting
things “wrong” in more complex scenarios if a wrong field is taken as
true, causing an error in a subsequent field. Then that gets cut out of
the diff or something. Maybe should make a note about that possibility.